+++ cmd/komodo/cmd/root.go
@@ -16,7 +16,7 @@ func init() {
 var rootCmd = &cobra.Command{
 	Use:   "komodo",
 	Short: "Komodo is a small wrapper around the LaTex build tool you preferred",
-	Long:  `TODO :)`,
+	Long:  `Komodo can be used to build LaTex Files and serve them to a webbrowser. You can even start a watcher so that you dont need to restart komodo and it automaticaly build you tex file on change.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		if len(args) == 0 {
 			cmd.Help()