+++ cmd/komodo/cmd/serve.go
@@ -1,0 +1,22 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+// serveCmd represents the serve command
+var serveCmd = &cobra.Command{
+	Use:   "serve",
+	Short: "Starts the webserver to serve the pdf file to a browser",
+	Long:  ``,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("serve called")
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(serveCmd)
+
+}